refactor(mobile): split agent activity pages
